refactor SKPathStroker
